crop_health_analyzer/requirements.txt
google-adk
uvicorn
google-cloud-storage
google-generativeai
python-dotenv
